Partial block writer clean-up and fix issue with incorrect handling of partial blocks
